修正乘法计算错误

+++ Numeric/Fixed32/Fixed32.Mul.cs
@@ -94,14 +94,34 @@
             var aint = a >> INTEGRAL_BITS;
             var bint = b >> INTEGRAL_BITS;
             // 小数部分
-            var afrac = (a & FRACTIONAL_MASK);
-            var bfrac = (b & FRACTIONAL_MASK);
+            var afrac = (ulong)(a & FRACTIONAL_MASK);
+            var bfrac = (ulong)(b & FRACTIONAL_MASK);
 
-            var rint  = (aint * bint) << FRACTIONAL_BITS; // 整数的积
+            var rint = (aint * bint) << FRACTIONAL_BITS; // 整数的积
             var rfrac = (afrac * bfrac) >> FRACTIONAL_BITS; // 小数的积
-            var rcrs  = (aint * bfrac + bint * afrac); // 交叉部分
+            var rcrs = (aint * (long)bfrac + bint * (long)afrac); // 交叉部分
 
-            return From(rint + rcrs + rfrac);
+            return From(rint + rcrs + (long)rfrac);
+            //var xl = x;
+            //var yl = y;
+
+            //var xlo = (ulong)(xl & 0x00000000FFFFFFFF);
+            //var xhi = xl >> FRACTIONAL_BITS;
+            //var ylo = (ulong)(yl & 0x00000000FFFFFFFF);
+            //var yhi = yl >> FRACTIONAL_BITS;
+
+            //var lolo = xlo * ylo;
+            //var lohi = (long)xlo * yhi;
+            //var hilo = xhi * (long)ylo;
+            //var hihi = xhi * yhi;
+
+            //var loResult = lolo >> FRACTIONAL_BITS;
+            //var midResult1 = lohi;
+            //var midResult2 = hilo;
+            //var hiResult = hihi << FRACTIONAL_BITS;
+
+            //var sum = (long)loResult + midResult1 + midResult2 + hiResult;
+            //return From(sum);
         }
     }
 }
